Extract search home page from perspective
