+++ src/main/scripts/src/app/core/search-home/search-home.component.scss
