+++ src/main/scripts/src/app/view/perspectives/search/all/search-all.component.html
