+++ src/main/scripts/src/app/view/perspectives/search/collections/search-collections.component.html
