+++ src/main/scripts/src/app/view/perspectives/search/documents/search-documents.component.ts
