+++ src/main/scripts/src/app/view/perspectives/search/links/search-links.component.html
