--- src/main/scripts/src/app/view/perspectives/search/results/search-results.component.html
@@ -1,49 +1,0 @@
-<ul class="nav nav-tabs">
-  <li role="presentation" class="nav-item">
-    <a [routerLink]="[viewPath()]"
-       [queryParams]="{query: query, perspective: 'search', searchTab: 'all'}"
-       routerLinkActive="active"
-       [routerLinkActiveOptions]="{exact: true}"
-       class="nav-link">
-      All
-    </a>
-  </li>
-  <li role="presentation" class="nav-item">
-    <a [routerLink]="[viewPath()]"
-       [queryParams]="{query: query, perspective: 'search', searchTab: 'collections'}"
-       routerLinkActive="active"
-       [routerLinkActiveOptions]="{exact: true}"
-       class="nav-link">
-      Collections
-    </a>
-  </li>
-  <li role="presentation" class="nav-item">
-    <a [routerLink]="[viewPath()]"
-       [queryParams]="{query: query, perspective: 'search', searchTab: 'documents'}"
-       routerLinkActive="active"
-       [routerLinkActiveOptions]="{exact: true}"
-       class="nav-link">
-      Documents
-    </a>
-  </li>
-  <li role="presentation" class="nav-item">
-    <a [routerLink]="[viewPath()]"
-       [queryParams]="{query: query, perspective: 'search', searchTab: 'links'}"
-       routerLinkActive="active"
-       [routerLinkActiveOptions]="{exact: true}"
-       class="nav-link">
-      Links
-    </a>
-  </li>
-  <li role="presentation" class="nav-item">
-    <a [routerLink]="[viewPath()]"
-       [queryParams]="{query: query, perspective: 'search', searchTab: 'views'}"
-       routerLinkActive="active"
-       [routerLinkActiveOptions]="{exact: true}"
-       class="nav-link">
-      Views
-    </a>
-  </li>
-</ul>
-
-<ng-template search-results></ng-template>
