+++ src/main/scripts/src/app/view/perspectives/search/search-results.directive.ts
