+++ src/main/scripts/src/app/view/perspectives/search/views/search-views.component.ts
